requirements.txt
APScheduler==3.6.3
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
decorator==5.1.1
executing==0.8.3
h11==0.13.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
lxml==4.9.0
matplotlib-inline==0.1.3
outcome==1.1.0
parso==0.8.3
persiantools==3.0.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tornado==6.2
traitlets==5.2.1.post0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5
wsproto==1.1.0
